+++ TrainingFramework/src/GameObject/MRectangle.h
@@ -5,26 +5,26 @@ class MRectangle
 {
 public:
 	MRectangle();
-	MRectangle(int x, int y, int height, int width);
+	MRectangle(float x, float y, int height, int width);
 	~MRectangle();
 
 	int isInteract(MRectangle anotherRec);
 	int caculDistanceToAnotherRecWithDirection(MRectangle anotherRec, int direction);
 
-	void setRecXY(int x , int y);
+	void setRecXY(float x , float y);
 	void setRecHeight(int height);
 	void setRecWidth(int width);
 
-	int getRecX();
-	int getRecY();
+	float getRecX();
+	float getRecY();
 	int getRecHeight();
 	int getRecWidth();
 
 	void toString();
 
 private:
-	int r_x;
-	int r_y;
+	float r_x;
+	float r_y;
 	int r_height;
 	int r_width;
 };
